refactor config
